Continued development on Navigation, Actions, and general optimizations.
added a non immediate stop for an action called halt() instead of the
immediate stop called abort()

added getters for robot objects such as lift and navX in robot.

added an interface for credibility of a sensors reading

added isAtSetpoint in superLift to help automate Actions utilizing lift.

moved calculations involving coordinates to another class.

removed AnglePID

removed distancControl PID in favor of static speed to waypoint. Also
added a boolean value for if the waypoint wants to be static speed or
variable.

-NEED TO-
make Waypoint and WaypointProvider not extensions of others and hold
hold their objects inside of themselves.